refactor: 로직 useComboBox로 분리
