+++ src/common/components/header/clientHeader/index.tsx
@@ -8,37 +8,21 @@ import Link from 'next/link';
 import mainlogo from '@/assets/mainlogo_wh.svg';
 import * as S from './clientHeader.css';
 import { useSearchClubInfinite } from '@/hooks/useSearchClubInfinite';
-import { useDebounce } from '@/hooks/useDebounce';
-import { useEffect, useState } from 'react';
 import InputCombobox from '../../inputCombobox';
 import SearchIcon from '@/assets/search.svg';
-import { useRouter } from 'next/navigation';
+import { useCombobox } from '@/hooks/useCombobox';
 
 function ClientHeader() {
-  const router = useRouter();
+  const {
+    debouncedSearch,
+    isComboBoxOpen,
+    setIsComboBoxOpen,
+    handleSearchChange,
+    handleNavigate,
+    handleKeyDown,
+  } = useCombobox();
   const isVisible = useScrollObserver();
-  const [searchdata, setSearchData] = useState('');
-  const debouncedSearch = useDebounce(searchdata);
-  const [isComboBoxOpen, setIsComboBoxOpen] = useState<boolean>(!!debouncedSearch);
-
   const { data: searchList } = useSearchClubInfinite({ debouncedSearch });
-  const handleSearchChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearchData(e.target.value);
-    console.log('Search Input Changed:', isComboBoxOpen);
-  };
-  useEffect(() => {
-    setIsComboBoxOpen(!!debouncedSearch);
-  }, [debouncedSearch]);
-
-  const handleNavigate = () => {
-    router.push(ROUTES.SEARCH(debouncedSearch));
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      handleNavigate();
-    }
-  };
 
   return (
     <Header isVisible={isVisible} className={S.InnerWrapper}>
